+++ webviz_subsurface/plugins/_parameter_analysis/_views/_parameter_response_view/_settings/_parameter_filter.py
@@ -25,4 +25,5 @@ class ParamRespParameterFilter(SettingsGroupABC):
             ].copy(),
             reset_on_ensemble_update=True,
             display_header=False,
+            include_sens_filter=True,
         ).layout